# Conflicts:
#	OsmAnd/res/values/strings.xml
#	OsmAnd/src/net/osmand/plus/settings/fragments/RouteLineAppearanceFragment.java

+++ OsmAnd/src/net/osmand/plus/track/Track3DStyle.kt
@@ -3,7 +3,7 @@ package net.osmand.plus.track
 class Track3DStyle(var visualizationType: Gpx3DVisualizationType,
 	var wallColorType: Gpx3DWallColorType,
 	var linePositionType: Gpx3DLinePositionType,
-	var additionalExaggeration: Int) {
+	var additionalExaggeration: Float) {
 
 	override fun equals(other: Any?): Boolean {
 		return super.equals(other) && other is Track3DStyle
